test(plugin-sdk): split recovery boundary coverage
