+++ src/plugin-sdk/session-store-runtime.test.ts
@@ -669,29 +669,6 @@ describe("session-store-runtime compatibility surface", () => {
     expect(loadInternalSessionEntry({ sessionKey, storePath })?.model).toBeUndefined();
   });
 
-  it("allows public recovery fields to change without an active core transaction", async () => {
-    const sessionKey = "agent:main:healthy-public-recovery";
-    await seedSessionEntry(sessionKey, {
-      sessionId: "healthy-session",
-      updatedAt: 10,
-    });
-
-    await patchSessionEntry({
-      sessionKey,
-      storePath,
-      update: () => ({
-        abortedLastRun: true,
-        restartRecoveryRuns: [{ lifecycleGeneration: "generation-1", runId: "run-1" }],
-      }),
-    });
-
-    expect(loadInternalSessionEntry({ sessionKey, storePath })).toMatchObject({
-      abortedLastRun: true,
-      restartRecoveryRuns: [{ lifecycleGeneration: "generation-1", runId: "run-1" }],
-      sessionId: "healthy-session",
-    });
-  });
-
   it("clears core recovery state when public replacements change session identity", async () => {
     const patchKey = "agent:main:telegram:direct:patch-rotation";
     const upsertKey = "agent:main:telegram:direct:upsert-rotation";
@@ -852,42 +829,6 @@ describe("session-store-runtime compatibility surface", () => {
     ).toBeUndefined();
     expect(
       loadInternalSessionEntry({ sessionKey: updateKey, storePath: updateStorePath }),
-    ).not.toHaveProperty("mainRestartRecovery");
-  });
-
-  it("rejects core recovery state from runtime-escaped creation inputs", async () => {
-    const mainRestartRecovery = {
-      chargedAttempts: 1,
-      cycleId: "cycle-injected",
-      revision: 1,
-    };
-    const patchSessionKey = "agent:main:patch-created";
-    await patchSessionEntry({
-      fallbackEntry: {
-        mainRestartRecovery,
-        sessionId: "patch-created",
-        updatedAt: 10,
-      } as unknown as SessionEntry,
-      sessionKey: patchSessionKey,
-      storePath,
-      update: () => ({ updatedAt: 20 }),
-    });
-    expect(loadInternalSessionEntry({ sessionKey: patchSessionKey, storePath })).not.toHaveProperty(
-      "mainRestartRecovery",
-    );
-
-    const upsertSessionKey = "agent:main:upsert-created";
-    await upsertSessionEntry({
-      entry: {
-        mainRestartRecovery,
-        sessionId: "upsert-created",
-        updatedAt: 10,
-      } as unknown as SessionEntry,
-      sessionKey: upsertSessionKey,
-      storePath,
-    });
-    expect(
-      loadInternalSessionEntry({ sessionKey: upsertSessionKey, storePath }),
     ).not.toHaveProperty("mainRestartRecovery");
   });
